server/requirements.txt
Flask
redis
pycryptodome
certifi
cffi
chardet
cryptography
Flask-Login==0.5.0
idna
MarkupSafe
Naked
pycparser
pyotp
python-dotenv
PyYAML
requests
robin-stocks==2.0.3
shellescape
urllib3
Werkzeug~=2.0.0
whitenoise
gunicorn
pymongo
flask_session